requirements.txt
supabase>=2.0.0
httpx>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
tenacity>=8.2.0
